Make test decorator work outside tests
Because some folks might be like me and might assume that they can call
`scan` on the modules in the `tests/fixtures` directory and might be
puzzled when they do this and get:

    AttributeError: 'Scanner' object has no attribute 'test'

This makes the decorator not barf when used outside the test suite.

I also added a `call_count` static attribute that might be useful in the
future for making assertions about how many times the `callback` is
called.

Fixes GH-46